split swift code to multiple files
