+++ flutter_vlc_player/ios/Classes/VlcViewBuilder.swift
@@ -1,0 +1,8 @@
+//
+//  VlcViewBuilder.swift
+//  flutter_vlc_player
+//
+//  Created by alireza setayesh on 3/2/21.
+//
+
+import Foundation
