+++ flutter_vlc_player/ios/Classes/VlcViewController.swift
@@ -1,0 +1,8 @@
+//
+//  VlcViewController.swift
+//  flutter_vlc_player
+//
+//  Created by alireza setayesh on 3/2/21.
+//
+
+import Foundation
